requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Phase 1 Requirements: Core Agent + Scheduling Advisor + Streamlit UI
langchain>=0.1.0
openai>=1.0.0
sqlalchemy>=2.0.23
streamlit>=1.28.0
python-dateutil>=2.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Common useful packages
requests==2.31.0
numpy==1.25.2
pandas==2.1.3

# Web framework (for future phases)
# flask==3.0.0
# fastapi==0.104.1
# uvicorn==0.24.0

# Database (now enabled for Phase 1)
psycopg2-binary==2.9.9

# ML/AI (for future phases)
# scikit-learn==1.3.2
# torch==2.1.1
# tensorflow==2.14.0 